fix: details fixed

+++ src/main/java/cl/nessfit/web/model/Request.java
@@ -1,15 +1,9 @@
 package cl.nessfit.web.model;
 
-import org.springframework.format.annotation.DateTimeFormat;
-
 import javax.persistence.*;
 import java.io.Serial;
 import java.io.Serializable;
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 import java.util.Set;
 
 @Entity
@@ -21,7 +15,7 @@ public class Request implements Serializable {
     @GeneratedValue (strategy = GenerationType.IDENTITY)
     private int id;
     private int status;
-    private int price;
+    private long price;
     private LocalDate register;
     private int quantity;
     @ManyToOne (fetch = FetchType.EAGER)
@@ -36,14 +30,14 @@ public class Request implements Serializable {
      * Method that gets the request price.
      * @return Request price.
      */
-    public int getPrice() {
+    public long getPrice() {
         return price;
     }
     /**
      * Method that set the request price.
      * @param price New installation price.
      */
-    public void setPrice(int price) {
+    public void setPrice(long price) {
         this.price = price;
     }
     /**
